chore: move auth helper to internal package
